requirements.txt
asgiref==3.5.2
atomicwrites==1.4.1
attrs==22.1.0
certifi==2022.6.15
charset-normalizer==2.1.0
colorama==0.4.5
coverage==6.4.2
DateTime==4.5
Django==4.1
django-environ==0.9.0
django-filter==22.1
django-phonenumber-field==6.3.0
django-pytest==0.2.0
django-rest-auth==0.9.5
djangorestframework==3.13.1
environ==1.0
gunicorn==20.1.0
idna==3.3
iniconfig==1.1.1
packaging==21.3
phonenumbers==8.12.52
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
pytest-cov==3.0.0
pytest-cover==3.0.0
pytest-coverage==0.0
pytz==2022.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
tzdata==2022.1
urllib3==1.26.11
whitenoise==6.2.0
zope.interface==5.4.0
